refactor(#792): decompose start_meter_polling, split meter_wiring by responsibility (sweep)
meter_wiring.rs 717 -> 406 (leaves the debt list). This is a real decomposition,
not a relocation: the ~310-line start_meter_polling closure is broken into a thin
timer body + refresh_chain_meter_row (one job — refresh a single chain's meter row
from the current engine readings). Both move to meter_wiring_poll.rs (timer/wiring
responsibility); the pure compute + signature helpers stay in meter_wiring.rs.
start_meter_polling is re-exported so crate::meter_wiring::start_meter_polling
(desktop_app) resolves. Behaviour-preserving — the #715 in-place VecModel reuse and
every per-field delta check are verbatim. adapter-gui lib tests (540) pass, zero warnings.
